plots! [RESULTS]
muscle-fatigue.png is the output plot for fatigue without rests. With a continuous decline in active fibers, there is then an increase in fatigued fibers. Constant contraction of the muscle limits recovery.

muscle-fatigue-resting.png is the output with rests. When rest periods are added, the active fibers gain more stability over time.